fix date of registration

+++ application/migrations/020_modify_admins_table.php
@@ -21,7 +21,13 @@ class Migration_modify_admins_table extends CI_Migration
 
     public function down()
     { 
-        $qry = $this->db->query("ALTER TABLE admin DROP COLUMN ord"); 
+       
+        $result = $this->db->query("SHOW COLUMNS FROM `admin` LIKE 'ord'");
+        // $exists = ($result)?TRUE:FALSE;
+        if($result->num_rows()>0)
+        {
+           $qry = $this->db->query("ALTER TABLE admin DROP COLUMN ord"); 
+        }
     }
 }
 ?>
